example/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
Django==4.2.2
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.11.2
gspot-django-auth==0.0.9
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
